Small fix to how the output of whereis is processed.
When I tried to run Jails on Linux I was hitting an assert, and it
turned out that when jai-linux wasn't found in path it wasn't removing
the prefix from the output because there was no space after the ':'.

Also, added a simple Zed debug configuration that allowed me to step
through jails in order to track this issue down.